backend/requirements.txt
django==4.2.11
django-rest-framework==0.1.0
djangorestframework==3.15.1
psycopg==3.1.18